+++ packages/pyramid_lint/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Assist: `convert_to_for_in_iterable_indexed_loop`.
+- Lint: `specify_icon_button_tooltip`.
 
 ## [2.0.3] - 2024-03-15